Extract some functionality to its own packages
Like variable and template handling, and logging
