+++ Taskvars.yml
@@ -5,6 +5,10 @@ GO_PACKAGES:
   .
   ./cmd/task
   ./internal/args
+  ./internal/compiler
+  ./internal/compiler/v1
   ./internal/execext
+  ./internal/logger
   ./internal/status
   ./internal/taskfile
+  ./internal/templater
